% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt_gene.R
\name{filt_gene}
\alias{filt_gene}
\title{Filter Gene Expression File}
\usage{
filt_gene(exp, filt, thresh = NULL)
}
\arguments{
\item{exp}{(Object) Name of expression data frame object generated using read_gene()}

\item{filt}{(Character) Filter option for additional filtering - either sum, var, or both}

\item{thresh}{Numeric value between 0 and 1 for variance filter with default 0.3}
}
\value{
Returns filtered gene expression data frame to global env and writes to working directory
}
\description{
A simple function for additional gene expression filtering
}
\examples{
\dontrun{
library(processPandaR)
expression <- read_gene('expression_test.csv','gene')
gen_hist(expression,'both')
filtered <- filt_gene(expression,'sum')
}
}
\author{
Walid Nashashibi (\url{https://github.com/wnashash/})
}
