% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_gene.R
\name{filter_gene}
\alias{filter_gene}
\title{Filter Gene Expression File}
\usage{
filter_gene(geneExp, filter, threshold = NULL)
}
\arguments{
\item{geneExp}{(Object) Name of expression data frame object generated using read_gene()}

\item{filter}{(Character) Filter option for additional filtering - either sum, var, or both}

\item{threshold}{Numeric value between 0 and 1 for variance filter with default 0.1}
}
\value{
Returns filtered gene expression data frame to global env and writes to working directory
}
\description{
A simple function for additional gene expression filtering
}
\examples{
\dontrun{
library(processNetZoo)
exp_path <- system.file("extdata", "expression_test.csv", package = "processNetZoo", mustWork = TRUE)
expression <- read_gene(exp_path,'gene')
generate_histogram(expression,'both')
filtered <- filter_gene(expression,'sum')
}
}
\author{
Walid Nashashibi (\url{https://github.com/wnashash/})
}
