% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_hist.R
\name{gen_hist}
\alias{gen_hist}
\title{Generate Histogram for Variance Visualization}
\usage{
gen_hist(exp, filt)
}
\arguments{
\item{exp}{(Object) Name of expression data frame object generated using read_gene()}

\item{filt}{(Character) Filter option for visualization - either sum, var, or both}
}
\value{
Returns histogram plot of row sums or row variance
}
\description{
A simple function to generate histogram of row sums and row variance
}
\examples{
\dontrun{
library(processPandaR)
expression <- read_gene('expression_test.csv','gene')
gen_hist(expression,'both')
}
}
\author{
Walid Nashashibi (\url{https://github.com/wnashash/})
}
