% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_histogram.R
\name{generate_histogram}
\alias{generate_histogram}
\title{Generate Histogram for Variance Visualization}
\usage{
generate_histogram(geneExp, filter)
}
\arguments{
\item{geneExp}{(Object) Name of expression data frame object generated using read_gene()}

\item{filter}{(Character) Filter option for visualization - either sum, var, or both}
}
\value{
Returns histogram plot of row sums or row variance or both
}
\description{
A simple function to generate histogram of row sums and row variance
}
\examples{
\dontrun{
library(processNetZoo)
expression <- read_gene('extdata/expression_test.csv','gene')
generate_histogram(expression,'sum')
}
}
\author{
Walid Nashashibi (\url{https://github.com/wnashash/})
}
