% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean.R
\name{read_clean}
\alias{read_clean}
\title{Clean Gene Expression File}
\usage{
read_clean(file, geneName)
}
\arguments{
\item{file}{(Character) Gene expression file path, including file extension}

\item{geneName}{(Character) Name used for gene feature column in gene expression data}
}
\value{
Returns clean gene expression data frame to global env and writes to wd
}
\description{
A simple function to read and clean gene expression data
}
\examples{
\dontrun{
library(processPandaR)
expression <- read_clean("expression.csv","gene")
}
}
\author{
Walid Nashashibi (\url{https://github.com/wnashash/})
}
