% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_pipelines.R
\name{run_pipeline}
\alias{run_pipeline}
\title{File Prep and Run Pipeline}
\usage{
run_pipeline(
  pipeline,
  fileGeneExp,
  colGeneExp,
  fileMetadata = NULL,
  colMetadata = NULL
)
}
\arguments{
\item{pipeline}{(Character) Choose between panda, lioness, pandaCondor, and pandaAlpaca}

\item{fileGeneExp}{(Character) File path of gene expression file}

\item{colGeneExp}{(Character) Name of gene feature column in expression file}

\item{fileMetadata}{(Character) File path of metadata file}

\item{colMetadata}{(Character) Name of treatment feature column in metadata file}
}
\value{
Returns output of selected pipeline to be used for visualization
}
\description{
A simple function to prepare data and run pipeline of choice
}
\examples{
\dontrun{
library(processPandaR)
alpacaResult <- run_pipeline("pandaAlpaca","expression_test.csv","gene","metadata.csv","RECURRENCE_ANY")
}
}
\author{
Walid Nashashibi (\url{https://github.com/wnashash/})
}
