% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_NetZoo.R
\name{select_NetZoo}
\alias{select_NetZoo}
\title{Select NetZooR Algorithm}
\usage{
select_NetZoo(alg, exp, base, comp = NULL)
}
\arguments{
\item{alg}{(Character) Choose between panda, lioness, condor, alpaca, and monster}

\item{exp}{(Object) Name of expression data frame object generated using read_gene() or filt_gene()}

\item{base}{(Character) Gene expression base group file path}

\item{comp}{(Character) Gene expression comparison group file path}
}
\value{
Returns output of selected algorithm to be used for visualization or second algorithm
}
\description{
A simple function to run NetZooR algorithm of choice
}
\details{
In order to run alpaca or monster, split_gene() must be run first
}
\examples{
\dontrun{
library(processNetZoo)
expression <- read_gene('expression_test.csv','gene')
design <- split_gene(expression,'metadata.csv','SUBJECT_ID','RECURRENCE_ANY')
result <- select_NetZoo('alpaca',expression,'expression_base.txt','expression_comp.txt')
}
}
\author{
Walid Nashashibi (\url{https://github.com/wnashash/})
}
