% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_gene.R
\name{split_gene}
\alias{split_gene}
\title{Split Gene Expression File Using Metadata}
\usage{
split_gene(exp, path, simpleID, treatment)
}
\arguments{
\item{exp}{(Object) Name of expression data frame object generated using read_gene() or filt_gene()}

\item{path}{(Character) Metadata file path}

\item{simpleID}{(Character) Metadata column with samples/IDs found in gene expression data}

\item{treatment}{(Character) Metadata column that designates control and treatment samples/IDs}
}
\value{
Returns design and writes control (base) and treatment (comp) to working directory
}
\description{
A simple function to split gene expression data, using metadata treatment column
}
\examples{
\dontrun{
library(processPandaR)
expression <- read_gene('expression_test.csv','gene')
gen_hist(expression,'both')
filtered <- filt_gene(expression,'sum')
design <- split_gene(filtered,'metadata.csv','SAMPLE_ID','RECURRENCE_ANY')
}
}
\author{
Walid Nashashibi (\url{https://github.com/wnashash/})
}
