% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_norm.R
\name{transform_norm}
\alias{transform_norm}
\title{Normalize Expression File}
\usage{
transform_norm(expData, method)
}
\arguments{
\item{expData}{(Object) Name of expression dataframe object generated using readClean()}

\item{method}{(Character) Either TMM, RLE, or UPQ normalization method from edgeR package}
}
\value{
Returns filtered, normalized expression dataframe
}
\description{
A simple function to normalize expression data and export
}
\details{
Data is normalized and exported to working directory for pandaPy()/lionessPy()
}
\examples{
\dontrun{
library(processPandaR)
normalized <- transformNorm(expression,"TMM")
}
}
\author{
Walid Nashashibi (\url{https://github.com/wnashash/})
}
