% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean.R
\name{read_clean}
\alias{read_clean}
\title{Filter Expression File}
\usage{
read_clean(file, geneName)
}
\arguments{
\item{file}{(Character) File path, including file extension}

\item{geneName}{(Character) Name used for gene feature in expression data}
}
\value{
Returns filtered, raw expression dataframe
}
\description{
A simple function to read and filter expression data
}
\details{
Data is then normalized and exported to working directory, using transformNorm()
}
\examples{
\dontrun{
library(processPandaR)
expression <- readClean("expression_data.csv","gene_name")
}
}
\author{
Walid Nashashibi (\url{https://github.com/wnashash/})
}
