% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gene.R
\name{read_gene}
\alias{read_gene}
\title{Read Gene Expression File}
\usage{
read_gene(path, geneName)
}
\arguments{
\item{path}{(Character) Gene expression file path, including file extension}

\item{geneName}{(Character) Name used for gene feature column in gene expression data}
}
\value{
Returns clean gene expression data frame to global env and writes to working directory
}
\description{
A simple function to read and clean gene expression data
}
\examples{
\dontrun{
library(processNetZoo)
expression <- read_gene('expression_test.csv','gene')
}
}
\author{
Walid Nashashibi (\url{https://github.com/wnashash/})
}
