% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gene.R
\name{read_gene}
\alias{read_gene}
\title{Read Gene Expression File}
\usage{
read_gene(path, gene_column, expr_file_out = "expression.txt")
}
\arguments{
\item{path}{(Character) Gene expression file path, including file extension}

\item{gene_column}{(Character) Name used for gene feature column in gene expression data}

\item{expr_file_out}{(Character) Optional argument to save output in working directory.  No
file will be saved when output_file is NULL.}
}
\value{
Returns clean gene expression data frame to global env and writes to working directory
}
\description{
A simple function to read and clean gene expression data
}
\examples{
\dontrun{
library(processNetZoo)
exp_path <- system.file("extdata",
     "expression_test.csv",
     package = "processNetZoo",
     mustWork = TRUE)

expression <- read_gene(exp_path,'gene')
}
}
\author{
Walid Nashashibi (\url{https://github.com/wnashash/})
}
