% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_NetZoo.R
\name{select_NetZoo}
\alias{select_NetZoo}
\title{Select NetZooR Algorithm}
\usage{
select_NetZoo(
  algorithm,
  geneExp,
  basePath,
  compPath = NULL,
  startSample = 1,
  endSample = "None"
)
}
\arguments{
\item{algorithm}{(Character) Choose between panda, lioness, condor, alpaca, and monster}

\item{geneExp}{(Object) Name of expression data frame object generated using read_gene() or filter_gene()}

\item{basePath}{(Character) Gene expression base group file path}

\item{compPath}{(Character) Gene expression comparison group file path}

\item{startSample}{(Numeric) Value to indicate lionessPy first sample - default is 1}

\item{endSample}{(Numeric) Value to indicate lionessPy last sample - default is 'None'}
}
\value{
Returns output of selected algorithm to be used for visualization or second algorithm
}
\description{
A simple function to run NetZooR algorithm of choice
}
\details{
In order to run alpaca or monster, split_gene() must be run first
}
\examples{
\dontrun{
library(processNetZoo)
exp_path <- system.file("extdata", "expression_test.csv", package = "processNetZoo", mustWork = TRUE)
expression <- read_gene(exp_path,'gene')
result <- select_NetZoo('lionessPy',expression,'expression.txt',startSample=11,endSample=20)
}
}
\author{
Walid Nashashibi (\url{https://github.com/wnashash/})
}
