% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_netZoo.R
\name{select_NetZoo}
\alias{select_NetZoo}
\title{Select NetZooR Algorithm}
\usage{
select_NetZoo(
  algorithm,
  expr_file = NULL,
  expr = NULL,
  motif_file = "default",
  ppi_file = "default",
  comp_file = NULL,
  start_sample = 1,
  end_sample = "None",
  design = NULL
)
}
\arguments{
\item{algorithm}{(Character) Choose between 'pandaPy', 'lionessPy', 'condor', 'alpaca', and 'monster'}

\item{expr_file}{(Character) Gene expression base group filepath}

\item{expr}{(Object) Name of expression data.frame object generated using `read_gene()` or `filter_gene()`}

\item{motif_file}{(Character) value indicating which motif data to use.  Options include NULL for Pearson correlation
matrix, 'default' for built-in data, or a user-specified filepath.}

\item{ppi_file}{(Character) value indicating which protein-protein interaction data to use.  Options include
NULL for no additional interaction, 'default' for built-in data, or a user-specified filepath.}

\item{comp_file}{(Character) Gene expression comparison group filepath for 'alpaca' networks}

\item{start_sample}{(Numeric) Value to indicate 'lionessPy' first sample - default is 1}

\item{end_sample}{(Numeric) Value to indicate 'lionessPy' last sample - default is 'None'}

\item{design}{(Numeric) Binary vector for 'monster' algorithm case control partition.}
}
\value{
Returns output of selected algorithm to be used for visualization or second algorithm
}
\description{
A simple function to run NetZooR algorithm of choice
}
\details{
In order to run alpaca or monster, split_gene() must be run first
}
\examples{
\dontrun{
library(processNetZoo)

# -- pandaPy --
exp_path <- system.file("extdata",
     "expr4_matched.txt",
     package = "netZooR",
     mustWork = TRUE)

result <- select_NetZoo(
     algorithm   = 'pandaPy',
     expr_file   = exp_path,
     motif_file  = NULL,
     ppi_file    = NULL
)


# -- lionessPy --
exp_path <- system.file("extdata",
     "expression_test.csv",
     package = "processNetZoo",
     mustWork = TRUE)

expression <- read_gene(exp_path,'gene')

result <- select_NetZoo(
     algorithm    = 'lionessPy',
     expr         = expression,
     expr_file    = 'expression.txt',
     start_sample = 11,
     end_sample   = 20
)
}
}
\author{
Walid Nashashibi (\url{https://github.com/wnashash/})
}
