% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_gene.R
\name{split_gene}
\alias{split_gene}
\title{Split Gene Expression File Using Metadata}
\usage{
split_gene(geneExp, metaPath, simpleID, treatment)
}
\arguments{
\item{geneExp}{(Object) Name of expression data frame object generated using read_gene() or filter_gene()}

\item{metaPath}{(Character) Metadata file path}

\item{simpleID}{(Character) Metadata column with samples/IDs found in gene expression data}

\item{treatment}{(Character) Metadata column that designates control and treatment samples/IDs}
}
\value{
Returns design and writes control (base) and treatment (comp) to working directory
}
\description{
A simple function to split gene expression data, using metadata treatment column
}
\examples{
\dontrun{
library(processNetZoo)
exp_path <- system.file("extdata", "expression_test.csv", package = "processNetZoo", mustWork = TRUE)
expression <- read_gene(exp_path,'gene')
generate_histogram(expression,'both')
filtered <- filter_gene(expression,'sum')
meta_path <- system.file("extdata", "metadata.csv", package = "processNetZoo", mustWork = TRUE)
design <- split_gene(filtered,meta_path,'SUBJECT_ID','RECURRENCE_ANY')
}
}
\author{
Walid Nashashibi (\url{https://github.com/wnashash/})
}
